Adjust makefile so that we don't provide a section address for BOOT.
This prevents the risk of a hex with code in bootloader segment. The only
code that uses this BOOT segment in the codebase is in the ASF NVM module,
in functions we don't use, but it's sadly in the same source file as
functions we do, so can't just exclude it from build so easily.